Converting to a bit of nunit/FakeItEasy to allow for easier testing
In order to feel better about RoundhousE I'd like to test the migration runners.
I wasn't able to do this easy using bdd. I'm creating a project for using NUnit
and FakeItEasy to allow me to more easily test the internals and to allow for
a bit higher code coverage around crucial functionality.