UI tweaks for better readability
I made some tweaks to the console text to make changes to the database more
easily identifiable. I also tried to guess if someone was trying to do a dry run
against a database that doesn't have the RoundhousE support tables created in it.
RoundhousE will give the user some information about running it agains WITHOUT
the dryrun switch.